Implement Frame.Context
